#1- Created The Full APP With Following Requirement
Banks Search Application

You need to develop a single page web app (preferably, but not necessarily in AngularJS).
The app should list and search for banks that are fetched from the API mentioned below. There should be a dropdown for cities (just put in 5 cities in there) and there is a search bar. As I type in the search area, the table should be dynamically filtered (client-side filtering). Search should be across all fields.

Here's the backend API to get a list of banks: https://vast-shore-74260.herokuapp.com/banks?city=MUMBAI

Essentials your applications should have:
bank search screen which would show a list of banks
user should be able to search by text for the banks, across all the fields (important: there would be no search button)
pagination for the results of the search, the user should be able to select page size
mark some banks as favourites. View banks that were marked as favourites (favourites should persist state event if the website is refreshed or reloaded)
API calls should be cached
your application should be deployed to Heroku / Github pages

For extra fun, you can try the following:
clicking on bank name would redirect to a bank page, with a route like banks/{bankid} that displays the details of the bank

Please note: below mock-ups are for reference only. You are free to do the styling as you please.

Deliverables:
Hosting URL on Heroku / Github pages
Github repo link to your solution
Time taken to complete this

Feel free to reach out to us if you have any questions.